requirements.txt
torch
torchvision
faiss-cpu
pillow
numpy
fastapi
uvicorn
python-multipart
pytest
